requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
openai==1.101.0
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
pydantic==2.9.2
streamlit==1.28.1
pandas==2.0.3
PyPDF2==3.0.1
pdfplumber==0.10.0
python-multipart==0.0.6
